Remove epoll back-off timeout
Hi!

Does anyone remember the reason for adding this sleep to the worker thread(s)?

I have for some time been trying to understand why there are a fair amount of discards in the jitter buffers even though the network looks fine (i.e. a steady RTP cadence). What I discovered is that my worker threads (8 to 16 depending on hardware) are spending lots of time sleeping on the job. The result is that the incoming socket buffers are filling up because there are not enough workers to dispatch packets up the stack in a timely manner, and when the workers polls the events they get multiple packets which will overflow the jitter buffers (set to handle 40 ms delay).

BR
Martin