pjsip-apps: Set initial log level to 1 (#2912)
When pjsua is started or when the python bindings are initialized
tons of debug log messages are emitted before the command line
"log-level" arguments are processed.  This causes quite a bit
of unnecessary log space to be used.  To combat this,
pj_log_set_level(1) is now called before anything else.  The
command line arguments can, of course, still set any level
later on.