Avoid assertions when using PJ_HAS_THREADS 0
This fixes assertions occuring when using PJSIP without thread support.
The assertions get triggered when `pj_thread_this` gets called.

This patch sets the `grp lock owner` pointer to a constant invalid
value. While this is ugly, the lock owner should not equal NULL while
locked and a constant seemed the simplest solution to achieve this.