Add additional unit test options
All of the unit tests will now accept a '-n/--no-trap' option
that will prevent the test's main.c wrapper from intercepting the
SIGSEGV and SIGABRT signals.  This allows the OS to generate
a coredump which can be examined by gdb.

The pjsip unit test wrapper will now accept a '-t/--tests' option
that will take a list of tests to run.  If you're working on
a specific area, this makes it easier to just test that area
without having to comment out the defines in test.h. The default
is to run all tests.

Added *bench*.htm to .gitignore.