sip_transport_tls: Allow wildcard certifcates
Rightly the use of wildcards in certificates is disallowed in accordance
with RFC5922. However, RFC2818 does make some allowances with regards to
their use when using subject alt names with DNS name types.

As such this patch creates a new setting for TLS transports called
'allow_wildcard_certs', which when enabled allows DNS name types, as
well as the common name that start with '*.' to match as a wildcard.

For instance: *.example.com
will match for: foo.example.com

Partial matching is not allowed, e.g. f*.example.com, foo.*.com, etc...
And the starting wildcard only matches for a single levels

For instance: *.example.com
will NOT match for: foo.bar.example.com

The new setting is disabled by default.