Using the correct policy when restarting the SRTP transport
When restarting the SRTP transport due to PJMEDIA_SRTP_CHECK_RTP_SEQ_ON_RESTART,
PJsip first stops the current SRTP transport to not maintain
the roll-over counter. The stop method cleans the tx_policy and
rx_policy variable, that are going to be used right after when trying to
start the transport again. The transport will not be able to start again
since the policies are not filled anymore and the SRTP transport will be
permanently disabled by bypass flag.

With this change, we are using the temporary policy that are held by the
SRTP structure and the transport will be capable of restart correctly..