Extend algorithm to find account for incoming calls in PJSUA(2)
With issue #1688, it is possible to set different request and
To-header URIs for outgoing SIP INVITEs. For incoming INVITEs,
where both URIs are SIP URIs, the algorithm that finds the corresponding
account considers the To header only:
If the username of the To-header is not equal to the account ID,
the first account of same transport/hostname or the default account is
used.

This algorithm is now extended to use the request URI's username as
lowest weight for the score. If the To header's username does not match
any account ID (local contact), the request URI's username may help.

Possible application: Similar to using a telephone URI in the To-header,
it might be a SIP URI containing a telephone number. In case of
telephone URI, PJSUA already uses the request URI.

Ref: RFC 3261, 8.2.2.1 To and Request-URI
   The To header field identifies the original recipient of the request
   designated by the user identified in the From field.  The original
   recipient may or may not be the UAS processing the request, due to
   call forwarding or other proxy operations.  A UAS MAY apply any
   policy it wishes to determine whether to accept requests when the To
   header field is not the identity of the UAS.  However, it is
   RECOMMENDED that a UAS accept requests even if they do not recognize
   the URI scheme (for example, a tel: URI) in the To header field, or
   if the To header field does not address a known or current user of
   this UAS.