Fix crash due to accessing deleted authentication session parent (#4331)
* Fix crash

that occurs when the app comes to the foreground and the IP connections were closed before, for example because the power saving mode was activated on the iPhone.
The reason was that the client authentication sessions of the regc object still pointed to a parent client authentication session that had already been deleted from the account.

* review remarks

using auth_clt_set_parent with a NULL parameter removes the parent binding
removed auth_session pointer that is not needed anymore