Avoid deadlock involving ioqueue read callback
Another approach to fix #4561.

The idea is to release the ioqueue-key lock before invoking the read callback. To maintain concurrency behavior, a callback queue is introduced, so if there is another thread invoking the callback, this thread will just queue the read event, that another thread will invoke the callback for this event.